+++ albatross_api/src/main/java/org/tokencreek/service/CompanyService.java
@@ -3,13 +3,8 @@ package org.tokencreek.service;
 
 import org.tokencreek.Company;
 
-import javax.decorator.Decorator;
-import javax.enterprise.inject.Default;
-import javax.enterprise.inject.Model;
-import javax.inject.Named;
-
 
 public interface CompanyService {
 
-    Company getCompanyById(Long companyId);
+    Company getCompanyById(Integer companyId);
 }
